requirements.txt
mysql-connector-python==8.2.0
pandas==2.1.4
scikit-learn==1.3.2
numpy==1.24.4